refactor: move pnpm logic into src/pnpm/ directory
Extract replacePnpmInnerPath into src/pnpm/replacePnpmInnerPath.ts
with a barrel export from src/pnpm/index.ts.
